requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.txt -o requirements.txt
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.4
    # via
    #   google-auth
    #   streamlit
certifi==2025.11.12
    # via
    #   pinecone
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via google-generativeai
google-auth==2.45.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-generativeai==0.8.6
    # via -r requirements.txt
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.3.0
    # via sqlalchemy
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.14.0
    # via altair
numpy==2.2.6
    # via
    #   pandas
    #   pydeck
    #   streamlit
orjson==3.11.5
    # via pinecone
packaging==24.2
    # via
    #   altair
    #   pinecone-plugin-assistant
    #   streamlit
pandas==2.3.3
    # via
    #   -r requirements.txt
    #   streamlit
pillow==12.0.0
    # via streamlit
pinecone==8.0.0
    # via -r requirements.txt
pinecone-plugin-assistant==3.0.1
    # via pinecone
pinecone-plugin-interface==0.0.7
    # via pinecone
proto-plus==1.27.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==22.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via google-generativeai
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pyodbc==5.3.0
    # via -r requirements.txt
pyparsing==3.2.5
    # via httplib2
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   pinecone
python-dotenv==1.2.1
    # via -r requirements.txt
pytz==2025.2
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.txt
    #   google-api-core
    #   pinecone-plugin-assistant
    #   streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sqlalchemy==2.0.45
    # via -r requirements.txt
streamlit==1.52.2
    # via -r requirements.txt
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   altair
    #   google-generativeai
    #   grpcio
    #   pinecone
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.2
    # via
    #   pinecone
    #   requests
watchdog==6.0.0
    # via streamlit
